+++ packages/draggable/foundation.ts
@@ -6,7 +6,7 @@ import {_DragEventDispatcher} from './dispatcher';
 import {AnimationHelper} from './utils/animation';
 import {Avatar} from './avatar';
 import {cssClasses, strings, numbers} from './constants';
-import {Axis} from './types';
+import {Axis, CustomScroll} from './types';
 
 /// Counter to generate a unique id for each instance.
 let idCounter: number = 0;
@@ -92,6 +92,10 @@ export class DraggableFoundation<D>  extends MDCFoundation<DraggableAdapter<D>> 
     /// [avatar] is a function to create a [DragAvatar] for this [Draggable].
     /// See [Draggable] constructor.
     avatar: Avatar;
+
+    // [customScroll] is function or boolean which allow to handle custom scrolling
+    // while dragging on Touch Devices
+    customScroll: CustomScroll = true;
 
     /// [touchAction] is string used as style for touch-action css property
     /// It is checked on dragStart to set given style to element